requirements.txt
librosa
tensorflow
numpy
paho-mqtt
tensorflow-hub
joblib
fastdtw
supabase